refactor http routes
