+++ backend/auth/google.go
@@ -1,4 +1,4 @@
-package user
+package auth
 
 import (
 	"context"
