+++ backend/server/http/handlers/organization.go
@@ -4,8 +4,6 @@ import (
 	"encoding/json"
 	"net/http"
 
-	"github.com/go-chi/chi/v5"
-
 	"github.com/trysourcetool/sourcetool/backend/organization"
 	"github.com/trysourcetool/sourcetool/backend/server/http/adapters"
 	"github.com/trysourcetool/sourcetool/backend/server/http/requests"
@@ -83,76 +81,6 @@ func (h *OrganizationHandler) CheckSubdomainAvailability(w http.ResponseWriter, 
 	response := &responses.StatusResponse{
 		Code:    http.StatusOK,
 		Message: "Subdomain is available",
-	}
-
-	if err := httputil.WriteJSON(w, http.StatusOK, response); err != nil {
-		httputil.WriteErrJSON(r.Context(), w, err)
-		return
-	}
-}
-
-// UpdateUser godoc
-// @ID update-organization-user
-// @Accept json
-// @Produce json
-// @Tags organizations
-// @Param userID path string true " "
-// @Param Body body requests.UpdateOrganizationUserRequest true " "
-// @Success 200 {object} responses.UpdateUserResponse
-// @Failure default {object} errdefs.Error
-// @Router /organizations/users/{userID} [put].
-func (h *OrganizationHandler) UpdateUser(w http.ResponseWriter, r *http.Request) {
-	req := requests.UpdateOrganizationUserRequest{
-		UserID: chi.URLParam(r, "userID"),
-	}
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		httputil.WriteErrJSON(r.Context(), w, err)
-		return
-	}
-
-	if err := httputil.ValidateRequest(req); err != nil {
-		httputil.WriteErrJSON(r.Context(), w, err)
-		return
-	}
-
-	out, err := h.service.UpdateUser(r.Context(), adapters.UpdateOrganizationUserRequestToDTOInput(req))
-	if err != nil {
-		httputil.WriteErrJSON(r.Context(), w, err)
-		return
-	}
-
-	if err := httputil.WriteJSON(w, http.StatusOK, adapters.UpdateOrganizationUserOutputToResponse(out)); err != nil {
-		httputil.WriteErrJSON(r.Context(), w, err)
-		return
-	}
-}
-
-// DeleteUser godoc
-// @ID delete-organization-user
-// @Accept json
-// @Produce json
-// @Tags organizations
-// @Param userID path string true " "
-// @Success 200 {object} responses.StatusResponse
-// @Failure default {object} errdefs.Error
-// @Router /organizations/users/{userID} [delete].
-func (h *OrganizationHandler) DeleteUser(w http.ResponseWriter, r *http.Request) {
-	req := requests.DeleteOrganizationUserRequest{
-		UserID: chi.URLParam(r, "userID"),
-	}
-	if err := httputil.ValidateRequest(req); err != nil {
-		httputil.WriteErrJSON(r.Context(), w, err)
-		return
-	}
-
-	if err := h.service.DeleteUser(r.Context(), adapters.DeleteOrganizationUserRequestToDTOInput(req)); err != nil {
-		httputil.WriteErrJSON(r.Context(), w, err)
-		return
-	}
-
-	response := &responses.StatusResponse{
-		Code:    http.StatusOK,
-		Message: "Successfully deleted user",
 	}
 
 	if err := httputil.WriteJSON(w, http.StatusOK, response); err != nil {
