+++ backend/server/http/requests/organization.go
@@ -7,13 +7,3 @@ type CreateOrganizationRequest struct {
 type CheckSubdomainAvailablityRequest struct {
 	Subdomain string `validate:"required"`
 }
-
-type UpdateOrganizationUserRequest struct {
-	UserID   string   `json:"-" validate:"required,uuid4"`
-	Role     *string  `json:"role" validate:"oneof=admin developer member"`
-	GroupIDs []string `json:"groupIds"`
-}
-
-type DeleteOrganizationUserRequest struct {
-	UserID string `param:"userID" validate:"required,uuid4"`
-}
