+++ backend/server/http/requests/user.go
@@ -1,93 +1,34 @@
 package requests
 
-type RequestMagicLinkRequest struct {
-	Email string `json:"email" validate:"required,email"`
-}
-
-type AuthenticateWithMagicLinkRequest struct {
-	Token     string `json:"token" validate:"required"`
-	FirstName string `json:"firstName"`
-	LastName  string `json:"lastName"`
-}
-
-type RefreshTokenRequest struct {
-	RefreshToken    string `validate:"required"`
-	XSRFTokenHeader string `validate:"required"`
-	XSRFTokenCookie string `validate:"required"`
-}
-
-type SaveAuthRequest struct {
-	Token string `json:"token" validate:"required"`
-}
-
-type InviteUsersRequest struct {
+type CreateUserInvitationsRequest struct {
 	Emails []string `json:"emails" validate:"required"`
 	Role   string   `json:"role" validate:"required,oneof=admin developer member"`
 }
 
-type GoogleOAuthCallbackRequest struct {
-	State string `validate:"required"`
-	Code  string `validate:"required"`
-}
-
-type UpdateUserRequest struct {
+type UpdateMeRequest struct {
 	FirstName *string `json:"firstName"`
 	LastName  *string `json:"lastName"`
 }
 
-type SendUpdateUserEmailInstructionsRequest struct {
+type SendUpdateMeEmailInstructionsRequest struct {
 	Email             string `json:"email" validate:"required,email"`
 	EmailConfirmation string `json:"emailConfirmation" validate:"required,email"`
 }
 
-type UpdateUserEmailRequest struct {
+type UpdateMeEmailRequest struct {
 	Token string `json:"token" validate:"required"`
 }
 
-type ResendInvitationRequest struct {
+type UpdateUserRequest struct {
+	UserID   string   `json:"-" validate:"required,uuid4"`
+	Role     *string  `json:"role" validate:"oneof=admin developer member"`
+	GroupIDs []string `json:"groupIds"`
+}
+
+type DeleteUserRequest struct {
+	UserID string `param:"userID" validate:"required,uuid4"`
+}
+
+type ResendUserInvitationRequest struct {
 	InvitationID string `json:"invitationId" validate:"required,uuid"`
-}
-
-type RegisterWithMagicLinkRequest struct {
-	Token     string `json:"token"`
-	FirstName string `json:"firstName"`
-	LastName  string `json:"lastName"`
-}
-
-type RequestInvitationMagicLinkRequest struct {
-	InvitationToken string `json:"invitationToken" validate:"required"`
-}
-
-type AuthenticateWithInvitationMagicLinkRequest struct {
-	Token string `json:"token" validate:"required"`
-}
-
-type RegisterWithInvitationMagicLinkRequest struct {
-	Token     string `json:"token" validate:"required"`
-	FirstName string `json:"firstName" validate:"required"`
-	LastName  string `json:"lastName" validate:"required"`
-}
-
-type AuthenticateWithGoogleRequest struct {
-	Code  string `json:"code" validate:"required"`
-	State string `json:"state" validate:"required"`
-}
-
-type RegisterWithGoogleRequest struct {
-	Token string `json:"token" validate:"required"`
-}
-
-type RequestInvitationGoogleAuthLinkRequest struct {
-	InvitationToken string `json:"invitationToken" validate:"required"`
-}
-
-type AuthenticateWithInvitationGoogleAuthLinkRequest struct {
-	Code  string `json:"code" validate:"required"`
-	State string `json:"state" validate:"required"`
-}
-
-type RegisterWithInvitationGoogleAuthLinkRequest struct {
-	Token     string `json:"token" validate:"required"`
-	FirstName string `json:"firstName" validate:"required"`
-	LastName  string `json:"lastName" validate:"required"`
 }
