+++ backend/server/http/responses/user.go
@@ -1,37 +1,7 @@
 package responses
 
-type RequestMagicLinkResponse struct {
-	Email string `json:"email"`
-	IsNew bool   `json:"isNew"`
-}
-
-type AuthenticateWithMagicLinkResponse struct {
-	AuthURL         string `json:"authUrl"`
-	Token           string `json:"token"`
-	HasOrganization bool   `json:"hasOrganization"`
-	IsNewUser       bool   `json:"isNewUser"`
-}
-
-type RefreshTokenResponse struct {
-	ExpiresAt string `json:"expiresAt"`
-}
-
-type SaveAuthResponse struct {
-	ExpiresAt   string `json:"expiresAt"`
-	RedirectURL string `json:"redirectUrl"`
-}
-
-type ObtainAuthTokenResponse struct {
-	AuthURL string `json:"authUrl"`
-	Token   string `json:"token"`
-}
-
-type InviteUsersResponse struct {
+type CreateUserInvitationsResponse struct {
 	UserInvitations []*UserInvitationResponse `json:"userInvitations"`
-}
-
-type SignOutResponse struct {
-	Domain string
 }
 
 type UserResponse struct {
@@ -60,15 +30,19 @@ type GetMeResponse struct {
 	User *UserResponse `json:"user"`
 }
 
+type UpdateMeResponse struct {
+	User *UserResponse `json:"user"`
+}
+
+type UpdateMeEmailResponse struct {
+	User *UserResponse `json:"user"`
+}
+
 type UpdateUserResponse struct {
 	User *UserResponse `json:"user"`
 }
 
-type UpdateUserEmailResponse struct {
-	User *UserResponse `json:"user"`
-}
-
-type ResendInvitationResponse struct {
+type ResendUserInvitationResponse struct {
 	UserInvitation *UserInvitationResponse `json:"userInvitation"`
 }
 
@@ -78,47 +52,4 @@ type UserGroupResponse struct {
 	GroupID   string `json:"groupId"`
 	CreatedAt string `json:"createdAt"`
 	UpdatedAt string `json:"updatedAt"`
-}
-
-type RequestInvitationMagicLinkResponse struct {
-	Email string `json:"email"`
-}
-
-type AuthenticateWithInvitationMagicLinkResponse struct {
-	AuthURL   string `json:"authUrl"`
-	Token     string `json:"token"`
-	IsNewUser bool   `json:"isNewUser"`
-}
-
-type RegisterWithMagicLinkResponse struct {
-	HasOrganization bool   `json:"hasOrganization"`
-	ExpiresAt       string `json:"expiresAt"`
-}
-
-type RegisterWithInvitationMagicLinkResponse struct {
-	ExpiresAt string `json:"expiresAt"`
-}
-
-type RequestGoogleAuthLinkResponse struct {
-	AuthURL string `json:"authUrl"`
-}
-
-type AuthenticateWithGoogleResponse struct {
-	FirstName                string `json:"firstName,omitempty"`
-	LastName                 string `json:"lastName,omitempty"`
-	AuthURL                  string `json:"authUrl"`
-	Token                    string `json:"token"`
-	HasOrganization          bool   `json:"hasOrganization"`
-	HasMultipleOrganizations bool   `json:"hasMultipleOrganizations"`
-	IsNewUser                bool   `json:"isNewUser"`
-}
-
-type RegisterWithGoogleResponse struct {
-	AuthURL         string `json:"authUrl"`
-	Token           string `json:"token"`
-	HasOrganization bool   `json:"hasOrganization"`
-}
-
-type RequestInvitationGoogleAuthLinkResponse struct {
-	AuthURL string `json:"authUrl"`
 }
